pyproject.toml
[project]
name = "nfcorpus-retrieval-agent"
version = "0.1.0"
description = "LLM-powered purple agent for NFCorpus document retrieval"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk[http-server]>=0.3.20",
    "uvicorn>=0.38.0",
    "pydantic-ai>=0.0.14",
    "openai>=1.0.0",
    "httpx>=0.28.1",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.1",
]
